+++ Sources/maccrabctl/Helpers.swift
@@ -145,8 +145,8 @@ extension MacCrabCtl {
     }
 
     static func printVersion() {
-        // LOCALIZE: "MacCrab Detection Engine v1.2.2"
-        print("MacCrab Detection Engine v1.2.2")
+        // LOCALIZE: "MacCrab Detection Engine v1.2.3"
+        print("MacCrab Detection Engine v1.2.3")
         // LOCALIZE: "License: Apache 2.0 (code), DRL 1.1 (rules)"
         print("License: Apache 2.0 (code), DRL 1.1 (rules)")
         print("https://github.com/maccrab-detection/maccrab")